feat: 전체 컨트롤러 float rod position 통일 및 팩토리 등록
- LLM 컨트롤러 rod position int→float 변경
- Planner float position 지원 및 range validation 강화
- Optimizer/Controller float type hints 통일
- PID/Proportional 컨트롤러 팩토리 등록
- blueprints.json initial_rod_position 보정 (rod scan 데이터 기반)
- 관련 테스트 업데이트

+++ src/controller/llm/optimizer.py
@@ -69,10 +69,10 @@ class BayesianOptimizer:
 
     def score(
         self,
-        positions: list[int],
-        current_position: int,
-        position_min: int = 0,
-        position_max: int = 228,
+        positions: list[float],
+        current_position: float,
+        position_min: float = 0.0,
+        position_max: float = 228.0,
     ) -> list[ScoredCandidate]:
         """목표 위치 후보에 안전 점수를 부여한다.
 
@@ -112,7 +112,7 @@ class BayesianOptimizer:
 
         return results
 
-    def refit(self, observations: list[tuple[int, float]]) -> None:
+    def refit(self, observations: list[tuple[float, float]]) -> None:
         """과거 관측 데이터로 GP 모델을 재피팅한다.
 
         Args:
